% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros
THis function takes two values x and y and sum them up}
\usage{
suma(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
the sum of x and y
}
\description{
Suma dos numeros
THis function takes two values x and y and sum them up
}
